requirements.txt
discord.py>=1.5,<2.0
aiohttp<4.0
aiosqlite>=0.16,<0.17